backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai
langchain
langchain-openai
langgraph
mangum
